Add tiny build helper warning.

+++ library/build.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if [ -z "$ANDROID_HOME" ]; then
+  echo 'You must have an $ANDROID_HOME environment variable set and Android 4.1 downloaded.'
+  exit 1
+fi
+
 rm -rf out NotificationCompat2-1.0.0.jar
 mkdir out
 javac -d out -cp "$ANDROID_HOME/platforms/android-16/android.jar" src/main/java/com/jakewharton/notificationcompat2/*
